refactor: split up large type and schema modules into smaller parts
