+++ frontend/src/lib/api/models/session.ts
@@ -1,0 +1,2 @@
+// Represents a unique ID of a session
+export type SessionId = string;
